+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI005 - 08/05/2023 - Add the CorrectDialog window.
 - LACLI005 - 08/05/2023 - Add the ErrorDialog window.
 - LACLI005 - 08/05/2023 - Add the QuestionDialog window.
+- LACLI008 - 13/05/2023 - Add the InformationWriter class.
 
 ### Changed
 - LACLI007 - 04/05/2023 - Change the formularys to apply new view.
@@ -22,7 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI005 - 09/05/2023 - Implements status dialogs.
 - LACLI005 - 10/05/2023 - Change ajust text in dialog.
 - LACLI003 - 11/05/2023 - Implements poweroff and reboot machine.
-- LACLI002 - 12/05/2023 - Change to string method of entity.
+- LACLI008 - 12/05/2023 - Change to string method of entity.
 
 ### Fixed
 - LACLI004 - 06/05/2023 - Selected rows bugs.